+++ internal/services/aws/rds/neptunegraph/graph_snapshots.go
@@ -8,7 +8,7 @@ import (
 	"vorpalstacks/internal/common/request"
 	"vorpalstacks/internal/core/logs"
 	storecommon "vorpalstacks/internal/store/aws/common"
-	ngstore "vorpalstacks/internal/store/aws/neptunegraph"
+	ngstore "vorpalstacks/internal/store/aws/rds/neptunegraph"
 )
 
 // CreateGraphSnapshot creates a point-in-time snapshot of the specified graph.