+++ internal/services/aws/rds/neptunegraph/procedure.go
@@ -11,7 +11,7 @@ import (
 	"vorpalstacks/internal/common/request"
 	"vorpalstacks/internal/core/logs"
 	"vorpalstacks/internal/core/storage/graphengine"
-	ngstore "vorpalstacks/internal/store/aws/neptunegraph"
+	ngstore "vorpalstacks/internal/store/aws/rds/neptunegraph"
 	"vorpalstacks/pkg/cypherparser"
 )