+++ internal/services/aws/rds/neptunegraph/service.go
@@ -20,7 +20,7 @@ import (
 	"vorpalstacks/internal/core/storage/graphengine"
 	"vorpalstacks/internal/eventbus"
 	storecommon "vorpalstacks/internal/store/aws/common"
-	ngstore "vorpalstacks/internal/store/aws/neptunegraph"
+	ngstore "vorpalstacks/internal/store/aws/rds/neptunegraph"
 	"vorpalstacks/internal/utils/aws/arn"
 )